crypto_project/requirements.txt
pytest
allure-pytest
pytest-xdist
requests
faker